bài viết - comment

+++ database/migrations/2022_10_03_130211_create_customers_table.php
@@ -16,7 +16,7 @@ class CreateCustomersTable extends Migration
         Schema::create('customers', function (Blueprint $table) {
             $table->increments("id");
             $table->string('name');
-            $table->decimal('phone');
+            $table->integer('phone');
             $table->string('email');
             $table->string('password');
             $table->string('avatar');
